+++ src/main/java/com/baro/auth/application/TokenDecrypter.java
@@ -1,4 +1,4 @@
-package com.baro.auth.infra.jwt;
+package com.baro.auth.application;
 
 public interface TokenDecrypter {
