+++ src/main/java/com/baro/auth/domain/Token.java
@@ -1,7 +1,6 @@
 package com.baro.auth.domain;
 
 public record Token(
-
         String accessToken,
         String refreshToken
 ) {
