+++ src/test/java/com/baro/auth/domain/AuthMemberTest.java
@@ -2,6 +2,7 @@ package com.baro.auth.domain;
 
 import org.junit.jupiter.api.Test;
 
+import static org.assertj.core.api.Assertions.assertThat;
 import static org.junit.jupiter.api.Assertions.*;
 
 @SuppressWarnings("NonAsciiCharacters")
@@ -9,6 +10,6 @@ class AuthMemberTest {
 
     @Test
     void AuthMember_동일성_테스트() {
-        assertEquals(new AuthMember(1000L), new AuthMember(1000L));
+        assertThat(new AuthMember(1000L)).isEqualTo(new AuthMember(1000L));
     }
 }
